add coredata model and custom classes

+++ FriendFaceChallenge/FriendFaceChallenge/DataModels/User.swift
@@ -12,7 +12,7 @@ struct User: Codable {
     var isActive: Bool
     
     var name: String
-    var age: Int
+    var age: Int16
     var company: String
     var email: String
     var address: String
